Clean: move sleep and think funcs to new file
